added(feat)[YARU-4]: support enum types
