+++ src/yaru/command.py
@@ -1,134 +1,14 @@
-import ast
 import inspect
 from argparse import (
     ArgumentParser,
-    BooleanOptionalAction,
     RawTextHelpFormatter,
     _SubParsersAction,
 )
-from dataclasses import dataclass
 from functools import partial, wraps
-from typing import Callable, Concatenate, Self, Sequence, get_args, overload
+from typing import Callable, Concatenate, Self, Sequence, overload
 
+from yaru.argument import CommandArgument
 from yaru.context import Context
-from yaru.exceptions import (
-    InvalidAnnotationTypeError,
-    InvalidArgumentTypeHintError,
-    MissingArgumentTypeHintError,
-)
-
-
-@dataclass
-class Arg:
-    """Used to provide extra metadata for an argument as part of an annotated type."""
-
-    help: str | None = None
-    metavar: str | tuple[str, ...] | None = None
-
-
-class _CommandArgument[T: Callable]:
-    """Used by the Command class to store the information concerning its cli arguments."""
-
-    class _Empty:
-        """Used to represent an argument without default value."""
-
-    def __init__(
-        self: Self,
-        name: str,
-        default: T | type[_Empty] = _Empty,
-        help: str | None = None,
-        metavar: str | tuple[str, ...] | None = None,
-    ) -> None:
-        self.name = name
-        self.default = default
-        self.help = help
-        self.metavar = metavar
-
-    @property
-    def is_optional(self) -> bool:
-        """`true` if the argument is optional, `false` otherwise."""
-        return self.default != self._Empty
-
-    @property
-    def arg_type(self) -> T:
-        """The type of this argument."""
-        return get_args(self.__orig_class__)[0]  # type: ignore
-
-    @staticmethod
-    def parse_literal_as_boolean(literal: str) -> bool:
-        """
-        Used by argparse when a non-optional boolean is used as a command argument.
-        It interprets strings like '1', '0', 'true', 'false' as booleans.
-        """
-        return bool(ast.literal_eval(literal))
-
-    @classmethod
-    def from_parameter(
-        cls: type[Self], parameter: inspect.Parameter
-    ) -> "_CommandArgument":
-        """Builds an instance of this class given a function parameter obtained by `inspect.signature`."""
-        if parameter.annotation == parameter.empty:
-            raise MissingArgumentTypeHintError(
-                f"'{parameter.name}' argument is not annotated, type hints are mandatory."
-            )
-
-        match get_args(parameter.annotation):
-            case ():
-                arg_type = parameter.annotation
-                metadata = Arg()
-            case (arg_type,):
-                metadata = Arg()
-            case (arg_type, metadata):
-                if not isinstance(metadata, Arg):
-                    raise InvalidAnnotationTypeError(
-                        f"'{parameter.name}' argument annotation must be of type {Arg}, found: {parameter.annotation}."
-                    )
-            case _:
-                raise InvalidArgumentTypeHintError(
-                    f"'{parameter.name}' argument couldn't be interpreted as a valid command argument, found: {parameter.annotation} ."
-                )
-
-        return _CommandArgument[arg_type](
-            parameter.name,
-            default=cls._Empty
-            if parameter.default == parameter.empty
-            else parameter.default,
-            help=metadata.help,
-            metavar=metadata.metavar,
-        )
-
-    def add_to_parser(self: Self, parser: ArgumentParser) -> None:
-        """Given a command parser, adds itself as an argument of that command."""
-        if self.arg_type is bool:
-            # Booleans are handled a tad different.
-            if self.is_optional:
-                parser.add_argument(
-                    f"--{self.name}",
-                    action=BooleanOptionalAction,
-                    default=self.default,
-                    help=self.help,
-                    metavar=self.metavar,
-                )
-            else:
-                parser.add_argument(
-                    self.name,
-                    type=self.parse_literal_as_boolean,
-                    help=self.help,
-                    metavar=self.metavar,
-                )
-        else:
-            if self.is_optional:
-                parser.add_argument(
-                    f"--{self.name}",
-                    type=self.arg_type,
-                    default=self.default,
-                    help=self.help,
-                    metavar=self.metavar,
-                )
-            else:
-                parser.add_argument(
-                    self.name, type=self.arg_type, help=self.help, metavar=self.metavar
-                )
 
 
 class Command:
@@ -140,7 +20,7 @@ class Command:
         self: Self,
         name: str,
         func: partial[Context],
-        arguments: Sequence[_CommandArgument] | None = None,
+        arguments: Sequence[CommandArgument] | None = None,
         help: str | None = None,
         aliases: Sequence[str] | None = None,
         prog: str | None = None,
@@ -173,11 +53,11 @@ class Command:
         return None
 
     @staticmethod
-    def parse_parameters(func: Callable) -> Sequence[_CommandArgument]:
+    def parse_parameters(func: Callable) -> Sequence[CommandArgument]:
         """Parses the command's arguments from the function's signature."""
         signature = inspect.signature(func)
         return [
-            _CommandArgument.from_parameter(p) for p in signature.parameters.values()
+            CommandArgument.from_parameter(p) for p in signature.parameters.values()
         ]
 
     @classmethod
